Move `MaxPolicy` to `range` package
