+++ model/src/main/kotlin/io/spine/validation/ValidationPlugin.kt
@@ -30,6 +30,7 @@ import io.spine.protodata.plugin.Plugin
 import io.spine.protodata.plugin.Policy
 import io.spine.protodata.plugin.ViewRepository
 import io.spine.protodata.render.Renderer
+import io.spine.validation.range.MaxPolicy
 import io.spine.validation.range.RangeFieldView
 import io.spine.validation.range.RangePolicy
 import io.spine.validation.required.RequiredFieldView
